requirements.txt
amqp==5.3.1
asgiref==3.8.1
attrs==22.2.0
billiard==4.2.1
celery==5.4.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
Django==5.1.5
django-admin-persian-fonts==0.2
django-cleanup==6.0.0
django-jalali==7.3.0
django-js-asset==3.1.0
django-render-partial==0.4
django-tinymce==4.1.0
django-widget-tweaks==1.5.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.2.2
drf-spectacular==0.25.1
ghasedak_sms==1.0.3
ghasedakpack==0.1.13
gunicorn==23.0.0
idna==3.10
inflection==0.5.1
jalali_core==1.0.0
jdatetime==5.2.0
jsonschema==4.17.3
kombu==5.4.2
packaging==24.2
persiantools==5.2.0
Pillow==11.1.0
prompt_toolkit==3.0.50
PyJWT==2.6.0
python-dateutil==2.9.0.post0
python-decouple==3.6
pytz==2025.1
PyYAML==6.0
redis==5.2.1
requests==2.32.3
six==1.17.0
sqlparse==0.5.3
tzdata==2025.1
uritemplate==4.1.1
urllib3==2.3.0
vine==5.1.0
wcwidth==0.2.13
whitenoise==6.9.0
